+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/CurrencyExchangeData.cs
@@ -2,5 +2,5 @@ namespace Poe2scout.Repositories.CurrencyExchange.Models;
 
 public record CurrencyExchangeData(
   int Epoch,
-  double Volume,
-  double MarketCap);
+  decimal Volume,
+  decimal MarketCap);
