+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/CurrencyExchangeHistory.cs
@@ -6,5 +6,5 @@ public record CurrencyExchangeHistory(
 
 public record CurrencyExchangeHistoryEntry(
   int Epoch,
-  float MarketCap,
-  float Volume);
+  decimal MarketCap,
+  decimal Volume);
