+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/CurrencyExchangeSnapshot.cs
@@ -5,5 +5,5 @@ public record CurrencyExchangeSnapshot(
   int LeagueId,
   int RealmId,
   List<CurrencyExchangeSnapshotPair> Pairs,
-  double Volume,
-  double MarketCap);
+  decimal Volume,
+  decimal MarketCap);
