+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/CurrencyExchangeSnapshotPair.cs
@@ -3,6 +3,6 @@ namespace Poe2scout.Repositories.CurrencyExchange.Models;
 public record CurrencyExchangeSnapshotPair(
   int CurrencyOneItemId,
   int CurrencyTwoItemId,
-  double Volume,
+  decimal Volume,
   CurrencyExchangeSnapshotPairData CurrencyOneData,
   CurrencyExchangeSnapshotPairData CurrencyTwoData);
