+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/CurrencyExchangeSnapshotPairData.cs
@@ -1,8 +1,8 @@
 namespace Poe2scout.Repositories.CurrencyExchange.Models;
 
 public record CurrencyExchangeSnapshotPairData(
-  double ValueTraded,
-  double RelativePrice,
-  int VolumeTraded,
-  int HighestStock,
-  double StockValue);
+  decimal ValueTraded,
+  decimal RelativePrice,
+  long VolumeTraded,
+  long HighestStock,
+  decimal StockValue);
