+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/PairData.cs
@@ -1,8 +1,8 @@
 namespace Poe2scout.Repositories.CurrencyExchange.Models;
 
 public record PairData(
-  double ValueTraded,
-  double RelativePrice,
-  double StockValue,
-  int VolumeTraded,
-  int HighestStock);
+  decimal ValueTraded,
+  decimal RelativePrice,
+  decimal StockValue,
+  long VolumeTraded,
+  long HighestStock);
