+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/PairHistoryEntryDataDetails.cs
@@ -2,9 +2,9 @@ namespace Poe2scout.Repositories.CurrencyExchange.Models;
 
 public record PairHistoryEntryDataDetails(
   int CurrencyItemId,
-  double ValueTraded,
-  double RelativePrice,
-  double StockValue,
-  int VolumeTraded,
-  int HighestStock);
+  decimal ValueTraded,
+  decimal RelativePrice,
+  decimal StockValue,
+  long VolumeTraded,
+  long HighestStock);
