+++ net/Poe2scout/Poe2scout/Repositories/CurrencyExchange/Models/SnapshotPair.cs
@@ -5,7 +5,7 @@ namespace Poe2scout.Repositories.CurrencyExchange.Models;
 public record SnapshotPair(
   int CurrencyExchangeSnapshotPairId,
   int CurrencyExchangeSnapshotId,
-  double Volume,
+  decimal Volume,
   Poe2scout.Models.CurrencyItem CurrencyOne,
   Poe2scout.Models.CurrencyItem CurrencyTwo,
   PairData CurrencyOneData,
